feat: handled all the errors in login and sign up page
First I built another reducer and action for errors and then I built a error component and put it inside of the app component.
And now whenever that we have an error we dispatch it.
Also whenever our error state has anything inside it, the error component will appear and show the error.