feat: put profie photo in header
When you log in, you'll go to the after login page and header will be there too.
And if you're logged in instead of login button you'll see your profile photo.
This works only if you log in with google and you'll see your google profile photo and if don't google
for log in you'll see a regular account icon.
This option isn't going to work for countries like Iran, Syriya, Cuba and North Korea
because the photo URL isn't available in those countries.