fix: added conditional routing base on user login condition
Now when user type an undefined url, the user will be redirected to the homepage.
Also when user is logged in and type login or signup url, the user will be redirected to the homepage.
Also when user is not logged in and type logout url, the user will be redirected to the homepage.
Also now we ahve a custom scroll bar.